Multi tagger's working (me thinks). Need the dishy timeout hack

+++ dis/MemmMultiTag.py
@@ -9,15 +9,11 @@ from lib.ml_framework import MachineLearningModule
 from lib.conllu import ConlluReader
 
 def setup(working_dir, fold_id): # executed on each node ONCE before jobs are scheduled
-	global tagger
 	from lib.MaxEntMarkovModel import MaxEntMarkovModel, Ratnaparkhi96Features, CollinsNormalisation
+	global tagger
 	tagger = MaxEntMarkovModel(feature_templates=Ratnaparkhi96Features, word_normaliser=CollinsNormalisation)
-	try:
-		tagger.load(working_dir, '-fold_%02d' % fold_id)
-	except:
-		return 1
-	else:
-		return 0
+	tagger.load(working_dir, '-fold_%02d' % fold_id)
+	return 0
 
 def cleanup():
 	import gc
@@ -26,24 +22,26 @@ def cleanup():
 
 def multi_tag(sentence):
 	s, _ = map(list, zip(*sentence))
-	# single = tagger.tag(s)
-	multi = tagger.multi_tag(s, 0.01)  # remove the very unlikely
-	# return single, multi
-	return {"".join(sentence): multi}
+	multi = tagger.multi_tag(s, 0.01)  # remove the very very unlikely
+	return {"".join(s): multi}
 
 def decompress_model(self, fold_id):
 	file_name = MaxEntMarkovModel.save_file(self.dir('working'), '-fold_%02d' % fold_id)
 	file_name_gz = file_name + ".gz"
-	if os.path.exists(file_name_gz):
-		print "Decompressing %s" % file_name_gz
-		with gzip.open(file_name_gz) as f_in:
-			data = pickle.load(f_in)
-			with open(file_name, 'wb') as f_out:
-				pickle.dump(data, f_out, -1)
-				print "Decompressed to %s" % file_name
-				return file_name
+	if not os.path.exists(file_name):
+		if os.path.exists(file_name_gz):
+			print "Decompressing %s" % file_name_gz
+			with gzip.open(file_name_gz) as f_in:
+				data = pickle.load(f_in)
+				with open(file_name, 'wb') as f_out:
+					pickle.dump(data, f_out, -1)
+					print "Decompressed to %s" % file_name
+					return file_name
+		else:
+			raise Exception("File don't exist, yo? %s" % file_name_gz)
 	else:
-		raise Exception("File don't exist, yo? %s" % file_name_gz)
+		print "File already decompressed?", file_name
+		return file_name
 
 class MemmMultiTag(MachineLearningModule):
 
@@ -58,19 +56,19 @@ class MemmMultiTag(MachineLearningModule):
 		num_folds = 10  # 10 fold cross validation
 		subset_size = len(training) / num_folds
 		http_server = None
+		reg = self.get('regularization')
 
 		# Tag the 1-left-out folds and accumulate for parser training.
 
 		for i in range(num_folds):
-			print "Fold", i
-			skip_fold = True if os.path.exists(backup_file_path(self.dir('working'), i, self.get('regularization'))+".gz", ) else False  # save time is we've done it already
+			skip_fold = True if os.path.exists(backup_file_path(self.dir('working'), i, reg)+".gz", ) else False  # save time is we've done it already
 			if not skip_fold:
 				current_model_file = decompress_model(self, i)  # unzip model
+				print "Fold", i, "from", i*subset_size, "to", subset_size
 				unsorted_tagging = training[i*subset_size:][:subset_size]
-				tagging = sorted(unsorted_tagging, key=lambda k: len(unsorted_tagging)) # sort to longest first
 				# learning = training[:i*subset_size] + training[(i+1)*subset_size:]
 				f = functools.partial(setup, self.dir('working'), i)  # make setup function with some parameters
-				cluster = dispy.SharedJobCluster(multi_tag, setup=f, cleanup=cleanup, reentrant=True)
+				cluster = dispy.JobCluster(multi_tag, setup=f, cleanup=cleanup, reentrant=True)
 
 				# Monitor cluster
 
@@ -82,7 +80,7 @@ class MemmMultiTag(MachineLearningModule):
 				# Create Jobs
 
 				jobs = []
-				for j, sentence in enumerate(tagging):
+				for j, sentence in enumerate(reversed(sorted(unsorted_tagging, key=lambda k: len(k)))):  # sort longest first?
 					job = cluster.submit(sentence)
 					job.id = (i+1) * (j+1)
 					jobs.append(job)
@@ -92,22 +90,24 @@ class MemmMultiTag(MachineLearningModule):
 				cluster.close()
 
 				# Collect the results in a single set.
-				faild = 0
+				failed = 0
+				print len(jobs), "jobs"
 				for job in jobs:
 					job()
 					if job.status != dispy.DispyJob.Finished:
-						faild += 1
+						failed += 1
 						# raise Exception('job %s failed: %s' % (job.id, job.exception))
 					else:
-						multi = job.result
-						self.tagged.update(multi)
-				print faild, "failed jobs"
+						print "Result", job.result
+						self.tagged.update(job.result)
+				print failed, "failed jobs"
 				self.save(self.tagged, i)
 				os.remove(current_model_file)  # remove unzipped version
 				# current_model_file = next_model_file
 			else:
 				continue
-		http_server.shutdown()
+		if http_server:
+			http_server.shutdown()
 		return False
 
 	def save(self, data, fold_id = 0):
